refactor: update state folder
